requirements.txt
cleverbot
discord.py
simplejson
praw